Move files to formatter sub directory
