+++ lib/cff.rb
@@ -27,9 +27,9 @@ require 'cff/entity'
 require 'cff/reference'
 require 'cff/model'
 require 'cff/file'
-require 'cff/formatter'
-require 'cff/apa_formatter'
-require 'cff/bibtex_formatter'
+require 'cff/formatter/formatter'
+require 'cff/formatter/apa_formatter'
+require 'cff/formatter/bibtex_formatter'
 
 # This library provides a Ruby interface to manipulate CITATION.cff files. The
 # primary entry points are Model and File.
