+++ lib/cff/formatter/apa_formatter.rb
