+++ lib/cff/formatter/bibtex_formatter.rb
