+++ lib/cff/formatter/formatter.rb
